Keep plugging away at register test
Need to return token from login, and then pass that into the command so that all
of our calls to register are as an authenticated admin user.

Think I just need to deal with the MimeRender instance, or more likely, use a
newtype wrapper with a ToJSON instance. Tomorrow thought, too tired.