Get parallel tests working
withTransactionSerializable can only retry when it catches the correct exception
from `postgres-simple`. `beam` uses LibPQ directly for the inserts we were
running, so never throws an exception indicating a serialization issue. As a
workaround, we implement our own retrying, that assumes an empty result set is
caused by a transaction conflict.